[risk=no] Remove our hacky API mapper and move all Controllers to the standard api package (#3710)
