+++ api/src/main/java/org/pmiops/workbench/interceptors/InterceptorUtils.java
@@ -1,23 +1,10 @@
 package org.pmiops.workbench.interceptors;
 
-import com.google.common.collect.ImmutableMap;
 import java.lang.reflect.Method;
-import java.util.Map;
 import java.util.regex.Pattern;
 import org.springframework.web.method.HandlerMethod;
 
 public class InterceptorUtils {
-
-  private static Map<String, String> apiImplMap =
-      ImmutableMap.of(
-          "org.pmiops.workbench.api.WorkspacesApiController",
-              "org.pmiops.workbench.workspaces.WorkspacesController",
-          "org.pmiops.workbench.api.BillingApiController",
-              "org.pmiops.workbench.billing.BillingController",
-          "org.pmiops.workbench.api.StatusAlertApiController",
-              "org.pmiops.workbench.statusalerts.StatusAlertController",
-          "org.pmiops.workbench.api.WorkspaceAdminApiController",
-              "org.pmiops.workbench.workspaceadmin.WorkspaceAdminController");
 
   private InterceptorUtils() {}
 
@@ -32,14 +19,8 @@ public class InterceptorUtils {
 
     // The matcher assumes that all Controllers are within the same package as the generated
     // ApiController (api package)
-    // The following code allows Controllers to be moved into other packages by specifying the
-    // mapping in `apiImplMap`
-    String controllerName;
-    if (apiImplMap.containsKey(apiControllerName)) {
-      controllerName = apiImplMap.get(apiControllerName);
-    } else {
-      controllerName = apiControllerPattern.matcher(apiControllerName).replaceAll("$1$2");
-    }
+    final String controllerName =
+        apiControllerPattern.matcher(apiControllerName).replaceAll("$1$2");
 
     Class<?> controllerClass;
     try {
